requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colored==2.2.4
currency-codes==23.6.4
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
tabulate==0.9.0
tomli==2.2.1
urllib3==2.2.3
